Split multiplication specs into separate files
